+++ example/lib/screens/complex_form.dart
@@ -180,13 +180,13 @@ class ComplexFormController extends AdvancedFormController {
       addSubform(humanSubform);
     } else {
       humanSubform.resetAll();
-      await removeSubform(humanSubform, close: false);
+      removeSubform(humanSubform, close: false);
     }
     if (type == SubformType.dog) {
       addSubform(dogSubform);
     } else {
       dogSubform.resetAll();
-      await removeSubform(dogSubform, close: false);
+      removeSubform(dogSubform, close: false);
     }
   }